requirements.txt
apache-airflow==2.7.1
pyspark==3.4.1
requests
pandas